Moving Commands to a separate folder. Only inter-platform calls shifted to commands now
