+++ packages/app-bridge/src/AppBridgeBase.ts
@@ -3,7 +3,6 @@
 import {
     Color,
     ColorPalette,
-    DispatchAction,
     Document,
     DocumentCategory,
     DocumentGroup,
@@ -14,34 +13,12 @@ import {
 } from './types';
 
 export interface AppBridgeBase {
-    dispatch(action: 'GetProjectId'): Promise<number>;
-    dispatch(action: 'GetEditorState'): boolean;
-    dispatch(action: 'GetTranslationLanguage'): string;
-    dispatch(action: 'GetColorPalettes'): Promise<ColorPalette[]>;
-    dispatch(action: DispatchAction, payload?: unknown): unknown;
-
-    /**
-     * @deprecated Use `dispatch('GetProjectId')` instead.
-     * This will be removed in version 4.0.0 of @frontify/app-bridge.
-     */
     getProjectId(): number;
 
-    /**
-     * @deprecated Use `dispatch('GetEditorState')` instead.
-     * This will be removed in version 4.0.0 of @frontify/app-bridge.
-     */
     getEditorState(): boolean;
 
-    /**
-     * @deprecated Use `dispatch('GetTranslationLanguage')` instead.
-     * This will be removed in version 4.0.0 of @frontify/app-bridge.
-     */
     getTranslationLanguage(): string;
 
-    /**
-     * @deprecated Use `await dispatch('GetColorPalettes')` instead.
-     * This will be removed in version 4.0.0 of @frontify/app-bridge.
-     */
     getColorPalettes(): Promise<ColorPalette[]>;
 
     getColorsByColorPaletteId(colorPaletteId: number): Promise<Color[]>;
