+++ packages/app-bridge/src/commands/AssetChooser.ts
@@ -1,0 +1,7 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { Command, CommandResponse } from '../types';
+
+export interface AssetChooser {
+    dispatch(command: Command.AssetChooser): CommandResponse[Command.AssetChooser];
+}
