+++ packages/app-bridge/src/commands/AssetViewer.ts
@@ -1,0 +1,7 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { Command, CommandResponse } from '../types';
+
+export interface AssetViewer {
+    dispatch(command: Command): CommandResponse[Command.AssetViewer];
+}
