+++ packages/app-bridge/src/types/Command.ts
@@ -1,0 +1,26 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { Asset } from './Asset';
+
+export enum Command {
+    AssetViewer = 'AssetViewer',
+    AssetChooser = 'assetChooser',
+}
+
+export enum CommandEvents {
+    AssetViewerOpened = 'assetViewerOpened',
+    AssetViewerClosed = 'assetViewerClosed',
+    AssetChooserOpened = 'assetChooserOpened',
+    AssetChooserClosed = 'assetChooserClosed',
+}
+
+export type CommandResponse = {
+    [Command.AssetViewer]: {
+        on(event: CommandEvents.AssetViewerOpened, callback: () => void): void;
+        on(event: CommandEvents.AssetViewerClosed, callback: () => void): void;
+    };
+    [Command.AssetChooser]: {
+        on(event: CommandEvents.AssetChooserOpened, callback: (selectedAssets: Asset[]) => void): void;
+        on(event: CommandEvents.AssetChooserClosed, callback: () => void): void;
+    };
+};
