--- packages/app-bridge/src/types/DispatchAction.ts
@@ -1,16 +1,0 @@
-/* (c) Copyright Frontify Ltd., all rights reserved. */
-
-export type DispatchAction =
-    | 'GetProjectId'
-    | 'GetEditorState'
-    | 'GetTranslationLanguage'
-    | 'GetColorPalettes'
-    | 'GetColorsByColorPaletteId'
-    | 'GetAllDocuments'
-    | 'GetDocumentGroups'
-    | 'GetDocumentPagesByDocumentId'
-    | 'GetDocumentCategoriesByDocumentId'
-    | 'GetUncategorizedPagesByDocumentId'
-    | 'GetDocumentSectionsByDocumentPageId'
-    | 'GetDocumentTargets'
-    | 'GetDocumentPageTargets';
